src/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.59.3"
python-dotenv = "^1.0.1"
numpy = "^2.2.1"
scikit-learn = "^1.6.0"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
